+++ send-transaction-service/src/lib.rs
@@ -1,7 +1,6 @@
 #![cfg(feature = "agave-unstable-api")]
 #![allow(clippy::arithmetic_side_effects)]
 pub mod send_transaction_service;
-pub mod send_transaction_service_next;
 pub mod send_transaction_service_stats;
 #[cfg(any(test, feature = "dev-context-only-utils"))]
 pub mod test_utils;
