+++ send-transaction-service/src/send_transaction_service.rs
@@ -5,7 +5,7 @@ use {
         },
         transaction_client::TransactionClient,
     },
-    crossbeam_channel::{Receiver, RecvTimeoutError},
+    crossbeam_channel::Receiver as CrossbeamReceiver,
     itertools::Itertools,
     log::*,
     solana_hash::Hash,
@@ -17,16 +17,15 @@ use {
     },
     solana_signature::Signature,
     std::{
-        collections::hash_map::{Entry, HashMap},
+        collections::{hash_map::Entry, HashMap},
         net::SocketAddr,
-        num::Saturating,
         sync::{
-            Arc, Mutex, RwLock,
             atomic::{AtomicBool, Ordering},
+            Arc, RwLock,
         },
-        thread::{self, Builder, JoinHandle, sleep},
         time::{Duration, Instant},
     },
+    tokio::{sync::mpsc, task::JoinHandle, time::sleep},
 };
 
 /// Maximum size of the transaction retry pool
@@ -57,12 +56,17 @@ const DEFAULT_BATCH_SEND_RATE_MS: u64 = 1;
 // The maximum transaction batch send rate in MS
 pub const MAX_BATCH_SEND_RATE_MS: usize = 100_000;
 
+/// The maximum duration the retry task may be configured to sleep before
+/// processing the transactions that need to be retried.
+pub const MAX_RETRY_SLEEP_MS: u64 = 1000;
+
 pub struct SendTransactionService {
-    receive_txn_thread: JoinHandle<()>,
-    retry_thread: JoinHandle<()>,
+    receive_txn_task: JoinHandle<()>,
+    retry_task: JoinHandle<()>,
     exit: Arc<AtomicBool>,
 }
 
+#[derive(Debug, Clone)]
 pub struct TransactionInfo {
     pub message_hash: Hash,
     pub signature: Signature,
@@ -152,55 +156,62 @@ impl Default for Config {
     }
 }
 
-/// The maximum duration the retry thread may be configured to sleep before
-/// processing the transactions that need to be retried.
-pub const MAX_RETRY_SLEEP_MS: u64 = 1000;
-
 impl SendTransactionService {
-    pub fn new<Client: TransactionClient + Clone + std::marker::Send + 'static>(
+    pub fn new<Client: TransactionClient + Clone + Send + 'static>(
         bank_forks: &Arc<RwLock<BankForks>>,
-        receiver: Receiver<TransactionInfo>,
+        crossbeam_receiver: CrossbeamReceiver<TransactionInfo>,
         client: Client,
         config: Config,
         exit: Arc<AtomicBool>,
     ) -> Self {
+        let (tokio_sender, tokio_receiver) = mpsc::channel(1000);
+
+        std::thread::spawn(move || {
+            while let Ok(transaction_info) = crossbeam_receiver.recv() {
+                if tokio_sender.blocking_send(transaction_info).is_err() {
+                    break;
+                }
+            }
+            debug!("Crossbeam-to-tokio bridge terminated");
+        });
+
         let stats_report = Arc::new(SendTransactionServiceStatsReport::default());
 
-        let retry_transactions = Arc::new(Mutex::new(HashMap::new()));
+        let (retry_sender, retry_receiver) = mpsc::unbounded_channel();
 
-        let receive_txn_thread = Self::receive_txn_thread(
-            receiver,
+        let receive_txn_task = Self::receive_txn_task(
+            tokio_receiver,
             client.clone(),
-            retry_transactions.clone(),
+            retry_sender,
             config.clone(),
             stats_report.clone(),
             exit.clone(),
         );
 
-        let retry_thread = Self::retry_thread(
+        let retry_task = Self::retry_task(
             bank_forks.clone(),
             client,
-            retry_transactions,
+            retry_receiver,
             config,
             stats_report,
             exit.clone(),
         );
+
         Self {
-            receive_txn_thread,
-            retry_thread,
+            receive_txn_task,
+            retry_task,
             exit,
         }
     }
 
-    /// Thread responsible for receiving transactions from RPC clients.
-    fn receive_txn_thread<Client: TransactionClient + std::marker::Send + 'static>(
-        receiver: Receiver<TransactionInfo>,
+    /// Task responsible for receiving transactions from RPC clients.
+    fn receive_txn_task<Client: TransactionClient + Send + 'static>(
+        mut receiver: mpsc::Receiver<TransactionInfo>,
         client: Client,
-        retry_transactions: Arc<Mutex<HashMap<Signature, TransactionInfo>>>,
+        retry_sender: mpsc::UnboundedSender<TransactionInfo>,
         Config {
             batch_send_rate_ms,
             batch_size,
-            retry_pool_max_size,
             default_max_retries,
             service_max_retries,
             ..
@@ -208,167 +219,209 @@ impl SendTransactionService {
         stats_report: Arc<SendTransactionServiceStatsReport>,
         exit: Arc<AtomicBool>,
     ) -> JoinHandle<()> {
-        let mut last_batch_sent = Instant::now();
-        let mut transactions = HashMap::new();
+        debug!("Starting send-transaction-service::receive_txn_task");
 
-        debug!("Starting send-transaction-service::receive_txn_thread");
-        Builder::new()
-            .name("solStxReceive".to_string())
-            .spawn(move || {
-                loop {
-                    let stats = &stats_report.stats;
-                    let recv_result =
-                        receiver.recv_timeout(Duration::from_millis(batch_send_rate_ms));
-                    if exit.load(Ordering::Relaxed) {
-                        break;
-                    }
-                    match recv_result {
-                        Err(RecvTimeoutError::Disconnected) => {
-                            info!("Terminating send-transaction-service.");
-                            exit.store(true, Ordering::Relaxed);
-                            break;
-                        }
-                        Err(RecvTimeoutError::Timeout) => {}
-                        Ok(transaction_info) => {
+        tokio::spawn(async move {
+            let mut transactions = HashMap::new();
+            let mut last_batch_sent = Instant::now();
+            let batch_send_rate = Duration::from_millis(batch_send_rate_ms);
+
+            loop {
+                if exit.load(Ordering::Relaxed) {
+                    break;
+                }
+                let stats = &stats_report.stats;
+
+                let elapsed = last_batch_sent.elapsed();
+                let timeout = batch_send_rate.saturating_sub(elapsed);
+
+                let should_send = if timeout.is_zero() {
+                    true
+                } else {
+                    match tokio::time::timeout(timeout, receiver.recv()).await {
+                        Ok(Some(transaction_info)) => {
                             stats.received_transactions.fetch_add(1, Ordering::Relaxed);
-                            let entry = transactions.entry(transaction_info.signature);
-                            let mut new_transaction = false;
-                            if let Entry::Vacant(_) = entry {
-                                if !retry_transactions
-                                    .lock()
-                                    .unwrap()
-                                    .contains_key(&transaction_info.signature)
-                                {
-                                    entry.or_insert(transaction_info);
-                                    new_transaction = true;
-                                }
-                            }
-                            if !new_transaction {
+
+                            if let Entry::Vacant(e) = transactions.entry(transaction_info.signature)
+                            {
+                                e.insert(transaction_info);
+                            } else {
                                 stats
                                     .received_duplicate_transactions
                                     .fetch_add(1, Ordering::Relaxed);
                             }
+
+                            transactions.len() >= batch_size
+                        }
+                        Ok(None) => {
+                            info!("Terminating send-transaction-service.");
+                            exit.store(true, Ordering::Relaxed);
+                            break;
+                        }
+                        Err(_) => true,
+                    }
+                };
+
+                if should_send && !transactions.is_empty() {
+                    stats
+                        .sent_transactions
+                        .fetch_add(transactions.len() as u64, Ordering::Relaxed);
+
+                    let wire_transactions = transactions
+                        .values()
+                        .map(|transaction_info| transaction_info.wire_transaction.clone())
+                        .collect::<Vec<Vec<u8>>>();
+
+                    client.send_transactions_in_batch(wire_transactions, stats);
+
+                    let last_sent_time = Instant::now();
+
+                    for (_signature, mut transaction_info) in transactions.drain() {
+                        let max_retries = transaction_info
+                            .get_max_retries(default_max_retries, service_max_retries);
+                        if max_retries == Some(0) {
+                            continue;
+                        }
+
+                        transaction_info.last_sent_time = Some(last_sent_time);
+
+                        if retry_sender.send(transaction_info).is_err() {
+                            break;
                         }
                     }
 
-                    if (!transactions.is_empty()
-                        && last_batch_sent.elapsed().as_millis() as u64 >= batch_send_rate_ms)
-                        || transactions.len() >= batch_size
-                    {
-                        stats
-                            .sent_transactions
-                            .fetch_add(transactions.len() as u64, Ordering::Relaxed);
-                        let wire_transactions = transactions
-                            .values()
-                            .map(|transaction_info| transaction_info.wire_transaction.clone())
-                            .collect::<Vec<Vec<u8>>>();
-                        client.send_transactions_in_batch(wire_transactions, stats);
-                        let last_sent_time = Instant::now();
-                        {
-                            // take a lock of retry_transactions and move the batch to the retry set.
-                            let mut retry_transactions = retry_transactions.lock().unwrap();
-                            let mut transactions_to_retry: usize = 0;
-                            let mut transactions_added_to_retry = Saturating::<usize>(0);
-                            for (signature, mut transaction_info) in transactions.drain() {
-                                // drop transactions with 0 max retries
-                                let max_retries = transaction_info
-                                    .get_max_retries(default_max_retries, service_max_retries);
-                                if max_retries == Some(0) {
-                                    continue;
-                                }
-                                transactions_to_retry += 1;
-
-                                let retry_len = retry_transactions.len();
-                                let entry = retry_transactions.entry(signature);
-                                if let Entry::Vacant(_) = entry {
-                                    if retry_len >= retry_pool_max_size {
-                                        break;
-                                    } else {
-                                        transaction_info.last_sent_time = Some(last_sent_time);
-                                        transactions_added_to_retry += 1;
-                                        entry.or_insert(transaction_info);
-                                    }
-                                }
-                            }
-                            let Saturating(retry_queue_overflow) =
-                                Saturating(transactions_to_retry) - transactions_added_to_retry;
-                            stats
-                                .retry_queue_overflow
-                                .fetch_add(retry_queue_overflow as u64, Ordering::Relaxed);
-                            stats
-                                .retry_queue_size
-                                .store(retry_transactions.len() as u64, Ordering::Relaxed);
-                        }
-                        last_batch_sent = Instant::now();
-                    }
+                    last_batch_sent = Instant::now();
                     stats_report.report();
                 }
-            })
-            .unwrap()
+            }
+
+            debug!("Exiting send-transaction-service::receive_txn_task");
+        })
     }
 
-    /// Thread responsible for retrying transactions
-    fn retry_thread<Client: TransactionClient + std::marker::Send + 'static>(
+    fn retry_task<Client: TransactionClient + Send + 'static>(
         bank_forks: Arc<RwLock<BankForks>>,
         client: Client,
-        retry_transactions: Arc<Mutex<HashMap<Signature, TransactionInfo>>>,
+        mut retry_receiver: mpsc::UnboundedReceiver<TransactionInfo>,
         config: Config,
         stats_report: Arc<SendTransactionServiceStatsReport>,
         exit: Arc<AtomicBool>,
     ) -> JoinHandle<()> {
-        debug!("Starting send-transaction-service::retry_thread.");
-        let sharable_banks = bank_forks.read().unwrap().sharable_banks();
-        let retry_interval_ms_default = MAX_RETRY_SLEEP_MS.min(config.retry_rate_ms);
-        let mut retry_interval_ms = retry_interval_ms_default;
-        Builder::new()
-            .name("solStxRetry".to_string())
-            .spawn(move || {
-                loop {
-                    sleep(Duration::from_millis(retry_interval_ms));
-                    if exit.load(Ordering::Relaxed) {
-                        break;
+        debug!("Starting send-transaction-service::retry_task.");
+
+        tokio::spawn(async move {
+            let sharable_banks = bank_forks.read().unwrap().sharable_banks();
+            let retry_interval_ms_default = MAX_RETRY_SLEEP_MS.min(config.retry_rate_ms);
+
+            let mut retry_transactions: HashMap<Signature, TransactionInfo> = HashMap::new();
+            let mut next_retry_time =
+                Instant::now() + Duration::from_millis(retry_interval_ms_default);
+
+            loop {
+                if exit.load(Ordering::Relaxed) {
+                    break;
+                }
+                let now = Instant::now();
+                let sleep_duration = if now >= next_retry_time {
+                    Duration::ZERO
+                } else {
+                    next_retry_time.duration_since(now)
+                };
+
+                tokio::select! {
+                    biased;
+
+                    msg = retry_receiver.recv() => {
+                        match msg {
+                            Some(transaction_info) => {
+                                let signature = transaction_info.signature;
+
+                                let stats = &stats_report.stats;
+
+                                if !retry_transactions.contains_key(&signature) {
+                                    let retry_len = retry_transactions.len();
+                                    if retry_len < config.retry_pool_max_size {
+                                        retry_transactions.insert(signature, transaction_info);
+                                        stats.retry_queue_size.store(
+                                            (retry_len + 1) as u64,
+                                            Ordering::Relaxed
+                                        );
+                                    } else {
+                                        stats.retry_queue_overflow.fetch_add(1, Ordering::Relaxed);
+                                    }
+                                }
+                                continue;
+                            }
+                            None => {
+                                break;
+                            }
+                        }
                     }
-                    let mut transactions = retry_transactions.lock().unwrap();
-                    if transactions.is_empty() {
-                        retry_interval_ms = retry_interval_ms_default;
-                    } else {
-                        let stats = &stats_report.stats;
-                        stats
-                            .retry_queue_size
-                            .store(transactions.len() as u64, Ordering::Relaxed);
 
-                        let BankPair {
-                            root_bank,
-                            working_bank,
-                        } = sharable_banks.load();
-                        let result = Self::process_transactions(
-                            &working_bank,
-                            &root_bank,
-                            &mut transactions,
-                            &client,
-                            &config,
-                            stats,
-                        );
-                        stats_report.report();
-
-                        // Adjust retry interval taking into account the time since the last send.
-                        retry_interval_ms = retry_interval_ms_default
-                            .checked_sub(
-                                result
-                                    .last_sent_time
-                                    .and_then(|last| Instant::now().checked_duration_since(last))
-                                    .and_then(|interval| interval.as_millis().try_into().ok())
-                                    .unwrap_or(0),
-                            )
-                            .unwrap_or(retry_interval_ms_default);
+                    _ = sleep(sleep_duration) => {
                     }
                 }
-            })
-            .unwrap()
+
+                while let Ok(transaction_info) = retry_receiver.try_recv() {
+                    let signature = transaction_info.signature;
+                    let stats = &stats_report.stats;
+
+                    if !retry_transactions.contains_key(&signature) {
+                        let retry_len = retry_transactions.len();
+                        if retry_len < config.retry_pool_max_size {
+                            retry_transactions.insert(signature, transaction_info);
+                            stats
+                                .retry_queue_size
+                                .store((retry_len + 1) as u64, Ordering::Relaxed);
+                        } else {
+                            stats.retry_queue_overflow.fetch_add(1, Ordering::Relaxed);
+                        }
+                    }
+                }
+
+                if retry_transactions.is_empty() {
+                    next_retry_time =
+                        Instant::now() + Duration::from_millis(retry_interval_ms_default);
+                } else {
+                    let stats = &stats_report.stats;
+                    stats
+                        .retry_queue_size
+                        .store(retry_transactions.len() as u64, Ordering::Relaxed);
+
+                    let BankPair {
+                        root_bank,
+                        working_bank,
+                    } = sharable_banks.load();
+
+                    let result = Self::process_transactions(
+                        &working_bank,
+                        &root_bank,
+                        &mut retry_transactions,
+                        &client,
+                        &config,
+                        stats,
+                    );
+                    stats_report.report();
+
+                    let retry_interval_ms = retry_interval_ms_default
+                        .checked_sub(
+                            result
+                                .last_sent_time
+                                .and_then(|last| Instant::now().checked_duration_since(last))
+                                .and_then(|interval| interval.as_millis().try_into().ok())
+                                .unwrap_or(0),
+                        )
+                        .unwrap_or(retry_interval_ms_default)
+                        .max(100);
+
+                    next_retry_time = Instant::now() + Duration::from_millis(retry_interval_ms);
+                }
+            }
+        })
     }
 
     /// Retry transactions sent before.
-    fn process_transactions<Client: TransactionClient + std::marker::Send + 'static>(
+    fn process_transactions<Client: TransactionClient + Send + 'static>(
         working_bank: &Bank,
         root_bank: &Bank,
         transactions: &mut HashMap<Signature, TransactionInfo>,
@@ -392,6 +445,7 @@ impl SendTransactionService {
             if transaction_info.durable_nonce_info.is_some() {
                 stats.nonced_transactions.fetch_add(1, Ordering::Relaxed);
             }
+
             if root_bank
                 .get_committed_transaction_status_and_slot(
                     &transaction_info.message_hash,
@@ -404,10 +458,12 @@ impl SendTransactionService {
                 stats.rooted_transactions.fetch_add(1, Ordering::Relaxed);
                 return false;
             }
+
             let signature_status = working_bank.get_committed_transaction_status_and_slot(
                 &transaction_info.message_hash,
                 &transaction_info.blockhash,
             );
+
             if let Some((nonce_pubkey, durable_nonce)) = transaction_info.durable_nonce_info {
                 let nonce_account = working_bank.get_account(&nonce_pubkey).unwrap_or_default();
                 let now = Instant::now();
@@ -425,6 +481,7 @@ impl SendTransactionService {
                     return false;
                 }
             }
+
             if transaction_info.last_valid_block_height < root_bank.block_height() {
                 info!("Dropping expired transaction: {signature}");
                 result.expired += 1;
@@ -454,11 +511,9 @@ impl SendTransactionService {
                         .and_then(|last| now.checked_duration_since(last))
                         .map(|elapsed| elapsed >= retry_rate)
                         .unwrap_or(true);
+
                     if need_send {
                         if transaction_info.last_sent_time.is_some() {
-                            // Transaction sent before is unknown to the working bank, it might have been
-                            // dropped or landed in another fork. Re-send it.
-
                             info!("Retrying transaction: {signature}");
                             result.retried += 1;
                             transaction_info.retries += 1;
@@ -501,7 +556,6 @@ impl SendTransactionService {
         stats.retries.fetch_add(result.retried, Ordering::Relaxed);
 
         if !batched_transactions.is_empty() {
-            // Processing the transactions in batch
             let wire_transactions = batched_transactions
                 .iter()
                 .filter_map(|signature| transactions.get(signature))
@@ -526,18 +580,20 @@ impl SendTransactionService {
         result
     }
 
-    pub fn join(self) -> thread::Result<()> {
-        self.receive_txn_thread.join()?;
+    pub async fn join(self) -> Result<(), tokio::task::JoinError> {
+        self.receive_txn_task.await?;
         self.exit.store(true, Ordering::Relaxed);
-        self.retry_thread.join()
+        self.retry_task.await?;
+        Ok(())
     }
 }
+
 #[cfg(test)]
 mod test {
     use {
         super::*,
         crate::test_utils::create_client_for_tests,
-        crossbeam_channel::{bounded, unbounded},
+        crossbeam_channel::unbounded,
         solana_account::AccountSharedData,
         solana_genesis_config::create_genesis_config,
         solana_nonce::{self as nonce, state::DurableNonce},
@@ -549,7 +605,30 @@ mod test {
         tokio::runtime::Handle,
     };
 
-    const GENESIS_LAMPORTS: u64 = 10_000_000_000;
+    #[tokio::test(flavor = "multi_thread")]
+    async fn crossbeam_bridge_compatibility() {
+        let bank = Bank::default_for_tests();
+        let bank_forks = BankForks::new_rw_arc(bank);
+        let (sender, receiver) = unbounded();
+
+        let client =
+            create_client_for_tests(Handle::current(), "127.0.0.1:0".parse().unwrap(), None, 1);
+
+        let send_transaction_service = SendTransactionService::new(
+            &bank_forks,
+            receiver,
+            client.clone(),
+            Config {
+                retry_rate_ms: 100,
+                ..Config::default()
+            },
+            Arc::new(AtomicBool::new(false)),
+        );
+
+        drop(sender);
+        send_transaction_service.join().await.unwrap();
+        client.cancel();
+    }
 
     #[tokio::test(flavor = "multi_thread")]
     async fn service_exit() {
@@ -565,14 +644,14 @@ mod test {
             receiver,
             client.clone(),
             Config {
-                retry_rate_ms: 1000,
+                retry_rate_ms: 100,
                 ..Config::default()
             },
             Arc::new(AtomicBool::new(false)),
         );
 
         drop(sender);
-        send_transaction_service.join().unwrap();
+        send_transaction_service.join().await.unwrap();
         client.cancel();
     }
 
@@ -580,7 +659,7 @@ mod test {
     async fn validator_exit() {
         let bank = Bank::default_for_tests();
         let bank_forks = BankForks::new_rw_arc(bank);
-        let (sender, receiver) = bounded(0);
+        let (sender, receiver) = unbounded();
 
         let dummy_tx_info = || TransactionInfo {
             message_hash: Hash::default(),
@@ -597,12 +676,13 @@ mod test {
         let exit = Arc::new(AtomicBool::new(false));
         let client =
             create_client_for_tests(Handle::current(), "127.0.0.1:0".parse().unwrap(), None, 1);
-        let _send_transaction_service = SendTransactionService::new(
+
+        let send_transaction_service = SendTransactionService::new(
             &bank_forks,
             receiver,
             client.clone(),
             Config {
-                retry_rate_ms: 1000,
+                retry_rate_ms: 100,
                 ..Config::default()
             },
             exit.clone(),
@@ -610,22 +690,31 @@ mod test {
 
         sender.send(dummy_tx_info()).unwrap();
 
-        thread::spawn(move || {
-            exit.store(true, Ordering::Relaxed);
-            client.cancel();
-        });
-
-        let mut option = Ok(());
-        while option.is_ok() {
-            option = sender.send(dummy_tx_info());
+        for _ in 0..5 {
+            let _ = sender.send(dummy_tx_info());
         }
+
+        exit.store(true, Ordering::Relaxed);
+
+        drop(sender);
+
+        let timeout_result =
+            tokio::time::timeout(Duration::from_secs(5), send_transaction_service.join()).await;
+
+        assert!(
+            timeout_result.is_ok(),
+            "Service did not exit within timeout"
+        );
+        timeout_result.unwrap().unwrap();
+
+        client.cancel();
     }
 
     #[tokio::test(flavor = "multi_thread")]
     async fn process_transactions() {
         agave_logger::setup();
 
-        let (mut genesis_config, mint_keypair) = create_genesis_config(GENESIS_LAMPORTS);
+        let (mut genesis_config, mint_keypair) = create_genesis_config(4);
         genesis_config.fee_rate_governor = solana_fee_calculator::FeeRateGovernor::new(0, 0);
         let (_, bank_forks) = Bank::new_with_bank_forks_for_tests(&genesis_config);
 
@@ -710,6 +799,7 @@ mod test {
             config.tpu_peers.clone(),
             leader_forward_count,
         );
+
         let result = SendTransactionService::process_transactions(
             &working_bank,
             &root_bank,
@@ -905,7 +995,7 @@ mod test {
     async fn retry_durable_nonce_transactions() {
         agave_logger::setup();
 
-        let (mut genesis_config, mint_keypair) = create_genesis_config(GENESIS_LAMPORTS);
+        let (mut genesis_config, mint_keypair) = create_genesis_config(4);
         genesis_config.fee_rate_governor = solana_fee_calculator::FeeRateGovernor::new(0, 0);
         let (_, bank_forks) = Bank::new_with_bank_forks_for_tests(&genesis_config);
         let leader_forward_count = 1;
@@ -1012,6 +1102,7 @@ mod test {
                 ..ProcessTransactionsResult::default()
             }
         );
+
         // Nonce expired case
         transactions.insert(
             rooted_signature,
@@ -1074,6 +1165,7 @@ mod test {
                 ..ProcessTransactionsResult::default()
             }
         );
+
         // ... or last_valid_block_height timeout has passed
         transactions.insert(
             Signature::default(),
@@ -1114,7 +1206,7 @@ mod test {
                 failed_transaction.message.recent_blockhash,
                 vec![],
                 last_valid_block_height,
-                Some((nonce_address, Hash::new_unique())), // runtime should advance nonce on failed transactions
+                Some((nonce_address, Hash::new_unique())),
                 None,
                 Some(Instant::now()),
             ),
@@ -1145,7 +1237,7 @@ mod test {
                 non_rooted_transaction.message.recent_blockhash,
                 vec![],
                 last_valid_block_height,
-                Some((nonce_address, Hash::new_unique())), // runtime advances nonce when transaction lands
+                Some((nonce_address, Hash::new_unique())),
                 None,
                 Some(Instant::now()),
             ),
@@ -1169,7 +1261,6 @@ mod test {
         transactions.clear();
 
         info!("Unknown durable-nonce transactions are retried until nonce advances...");
-        // simulate there was a nonce transaction sent 4 seconds ago (> the retry rate which is 2 seconds)
         transactions.insert(
             Signature::default(),
             TransactionInfo::new(
@@ -1199,8 +1290,8 @@ mod test {
                 ..ProcessTransactionsResult::default()
             }
         );
-        // Advance nonce, simulate the transaction was again last sent 4 seconds ago.
-        // This time the transaction should have been dropped.
+
+        // Advance nonce
         for transaction in transactions.values_mut() {
             transaction.last_sent_time = Some(Instant::now().sub(Duration::from_millis(4000)));
         }
